Refactor server structure
