+++ cmd/logging.go
@@ -1,0 +1,20 @@
+package main
+
+import (
+	"io"
+	"log"
+	"os"
+)
+
+func setupLogging(path string) (io.Closer, error) {
+	logFile, err := os.OpenFile(path, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
+	if err != nil {
+		return nil, err
+	}
+
+	mw := io.MultiWriter(os.Stdout, logFile)
+	log.SetOutput(mw)
+	log.SetFlags(log.Ldate | log.Ltime | log.Lshortfile)
+
+	return logFile, nil
+}
